requirements.txt
requests~=2.22.0
discord.py==1.6.0